requirements.txt
asgiref==3.7.2
attrs==23.1.0
boto3==1.35.22
botocore==1.35.22
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==43.0.1
distlib==0.3.2
Django==4.0
django-cors-headers==4.3.1
django-redirect==0.2
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dynaconf==3.2.6
filelock==3.0.12
Flask==2.2.3
gTTS==2.3.2
h11==0.14.0
hash-calc==1.1.0
hvac==2.3.0
idna==3.4
importlib-metadata==6.5.1
include==0.2.2
itsdangerous==2.1.2
Jinja2==3.0.1
jmespath==1.0.1
knox==0.1.14
loguru==0.7.2
MarkupSafe==2.1.2
mysqlclient==2.1.1
numpy==1.24.3
opencv-python==4.8.1.78
outcome==1.2.0
packaging==23.1
Pillow==9.5.0
platformdirs==2.0.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
pypng==0.20220715.0
PyQRCode==1.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==5.4.1
requests==2.28.2
s3transfer==0.10.2
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.2
stripe==10.11.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.9.0
tzdata==2021.5
urllib3==1.26.15
validators==0.34.0
virtualenv==20.6.0
webdriver-manager==3.8.6
Werkzeug==2.2.3
win32-setctime==1.1.0
wsproto==1.2.0
zipp==3.15.0
